Upload of SG-Lib 4.9.1 fro Matlab R2020a
Release 4.9.1 has been adapted for R2020a and is strongly recommended as the latest version. In version 2020a the programming interface of SimScape is modified ("File Solid" instead of "Solid" or changed "Default Names" of the "Connectors" of "Subsystems" etc.). Also callback functions of the user interface seem to react differently. SG-Lib 4.9.1 is valid for 400 days.
The function “loadweb" now links to the changed directory of the WEB-Server of the TU München. Release 4.9.1 does not have mat files included. Please use the laodweb function. Visit SG-Lib.org for any additional information